requirements.txt
npm==0.1.1
optional-django==0.1.0
Pillow==8.3.2
PyMySQL==1.0.2
